Input: xpad: Fix the usage of usb_control_msg()
Passing a NULL "data" and a non zero "length" is incorrect and result
in an invalid pointer being programmed into corresponding transfer
descriptor resulting in a USB controller halt. Since usb_control_msg
can't accept buffers allocated on stack, switch to using
usb_control_msge_recv() helper that would do the allocation for
us. This also fixes incorrect usage of usb_sndctrlpipe() (this is a
USB_DIR_IN transfer, so it should be usb_rcvctrlpipe())

Some third-party controllers, such as the HORPIAD FPS for Nintendo Switch and
Gamesir-G3w, require a specific packet that the first-party XInput driver sends
before it will start sending reports. It's not currently known what this packet
does, but since the first-party driver always sends it's unlikely that this
could cause issues with existing controllers.

Fixes: add53ac5e235 ("Input: xpad - add support for HORIPAD FPS for Nintendo Switch")
Co-authored-by: Vicki Pfau <vi@endrift.com>
Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>
[squashed 0becf9bd35a2daceb1d190a5988c70ec65bf2959 & d20eff6e1c6735e5ddf361108a2ae9e708fc8bda]
Signed-off-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>